backend/pyproject.toml
[project]
name = "flowvoice-backend"
version = "0.1.0"
description = "FlowVoice - Real-time AI speech accessibility companion"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
